fire-detection-monitoring/exporter/requirements.txt
prometheus_client==0.19.0
paho-mqtt==1.6.1
aiohttp==3.9.1
Pillow==10.2.0
python-dotenv==1.0.0
uvicorn==0.25.0
fastapi==0.109.0
requests==2.31.0